fix: disable loading animation by default to prevent issues with npx
Running build-sizes via npx was only outputting half of the metrics, and
they were in a different order. I think it has something to do with
the ANSI escape codes used by the loading animation. build-sizes would
work correctly when installing the package globally or to a project.

The loading animation is now disabled by default and can be enabled with
the "--loader" or "-l" flag.